BREAKING CHANGE: Update tests
